+++ src/org/interworldtransport/cladosGTest/CoreMonadRealDTest.java
@@ -134,7 +134,7 @@ class CoreMonadRealDTest {
 
 	@Test
 	public void testBiMathOps2() throws FieldBinaryException, CladosMonadBinaryException, CladosMonadException {
-		tM8.gradePart((short) 4).normalize();
+		tM8.gradePart((byte) 4).normalize();
 
 		tM6.multiplyLeft(tM8).dualLeft();
 		tM6.scale(new RealD(tM6.getCoeff((short) 0), -1f));
