+++ src/org/interworldtransport/cladosGTest/CoreMonadRealFTest.java
@@ -146,7 +146,7 @@ class CoreMonadRealFTest {
 	public void testBiMathOps2()
 			throws FieldBinaryException, FieldException, CladosMonadBinaryException, CladosMonadException {
 
-		tM8.gradePart((short) 4).normalize();
+		tM8.gradePart((byte) 4).normalize();
 		tM6.multiplyLeft(tM8).dualLeft();
 		tM6.scale(new RealF(tM6.getCoeff((short) 0), -1f));
 		assertTrue(tM6.isGEqual(tM7));
